requirements.txt
websockets
requests
pillow